LPS-59589 move common code to common class or jsp to avoid duplicated code
