+++ modules/frontend/frontend-taglib/src/META-INF/resources/card/horizontal_card/init.jsp
@@ -1,0 +1,22 @@
+<%--
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+--%>
+
+<%@ include file="/card/init.jsp" %>
+
+<%
+String icon = (String)request.getAttribute("liferay-frontend:card:icon");
+String text = (String)request.getAttribute("liferay-frontend:card:text");
+%>
