+++ modules/frontend/frontend-taglib/src/META-INF/resources/card/horizontal_card/page.jsp
@@ -14,7 +14,7 @@
  */
 --%>
 
-<%@ include file="/horizontal_card/init.jsp" %>
+<%@ include file="/card/horizontal_card/init.jsp" %>
 
 <div class="card-horizontal <%= Validator.isNotNull(cssClass) ? cssClass : StringPool.BLANK %>" <%= AUIUtil.buildData(data) %> >
 	<div class="card-row card-row-padded <%= showCheckbox ? "selectable" : StringPool.BLANK %>">
